modified setup and cleanup scripts

+++ scripts/cleanup.sh
@@ -3,8 +3,10 @@ set -euxo pipefail
 
 source ~/.${INSTRUQT_PARTICIPANT_ID}-env.sh
 
-cd ${TF_BASE}/boundary-demo-okta
-terraform destroy -auto-approve
+if [[ "$TF_VAR_use_okta" == "true"]]; then
+    cd ${TF_BASE}/boundary-demo-okta
+    terraform destroy -auto-approve
+fi
 
 cd ${TF_BASE}/boundary-demo-init
 terraform destroy -auto-approve
